Safari-611.1.10/LayoutTests/media/modern-media-controls/start-support/start-support-fullscreen-expected.txt

Testing the StartSupport behavior with no source.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Media entered fullscreen
PASS mediaController.controls.showsStartButton is false

PASS successfullyParsed is true

TEST COMPLETE
